+++ src/internal/interpreter/map-interpreter.errors.ts
@@ -65,7 +65,7 @@ export class MapInterpreterErrorBase extends ErrorBase {
       `${this.kind}: ${this.message}`,
       this.astPath ? `AST Path: ${this.astPath.join('.')}` : undefined,
       this.metadata?.node?.location
-        ? `Original Map Location: Line ${this.metadata.node.location.line}, column ${this.metadata.node.location.column}`
+        ? `Original Map Location: Line ${this.metadata.node.location.start.line}, column ${this.metadata.node.location.start.column}`
         : undefined,
     ]
       .filter(line => line !== undefined && line !== '')
@@ -119,7 +119,7 @@ export class MappedHTTPError<T> extends HTTPError {
         : undefined,
       this.astPath ? `AST Path: ${this.astPath.join('.')}` : undefined,
       this.metadata?.node?.location
-        ? `Original Map Location: Line ${this.metadata.node.location.line}, column ${this.metadata.node.location.column}`
+        ? `Original Map Location: Line ${this.metadata.node.location.start.line}, column ${this.metadata.node.location.start.column}`
         : undefined,
     ]
       .filter(line => line !== undefined && line !== '')
@@ -142,7 +142,7 @@ export class JessieError extends MapInterpreterErrorBase {
       this.originalError.toString(),
       this.astPath ? `AST Path: ${this.astPath.join('.')}` : undefined,
       this.metadata?.node?.location
-        ? `Original Map Location: Line ${this.metadata.node.location.line}, column ${this.metadata.node.location.column}`
+        ? `Original Map Location: Line ${this.metadata.node.location.start.line}, column ${this.metadata.node.location.start.column}`
         : undefined,
     ]
       .filter(line => line !== undefined && line !== '')
@@ -167,7 +167,7 @@ export class MappedError<T> extends MapInterpreterErrorBase {
         : undefined,
       this.astPath ? `AST Path: ${this.astPath.join('.')}` : undefined,
       this.metadata?.node?.location
-        ? `Original Map Location: Line ${this.metadata.node.location.line}, column ${this.metadata.node.location.column}`
+        ? `Original Map Location: Line ${this.metadata.node.location.start.line}, column ${this.metadata.node.location.start.column}`
         : undefined,
     ]
       .filter(line => !!line)