+++ src/internal/parser.ts
@@ -1,16 +1,16 @@
 import {
-  assertMapDocumentNode,
-  assertProfileDocumentNode,
   EXTENSIONS,
+  isMapDocumentNode,
+  isProfileDocumentNode,
   MapDocumentNode,
   ProfileDocumentNode,
 } from '@superfaceai/ast';
 import { parseMap, parseProfile, Source } from '@superfaceai/parser';
-import { createHash } from 'crypto';
 import { promises as fsp } from 'fs';
 import { join as joinPath } from 'path';
 
 import { Config } from '../config';
+import { UnexpectedError } from '.';
 
 export class Parser {
   private static mapCache: Record<string, MapDocumentNode> = {};
@@ -25,45 +25,45 @@ export class Parser {
       scope?: string;
     }
   ): Promise<MapDocumentNode> {
-    const hash = Parser.hash(input);
-    const hashedName = `${info.providerName}-${hash}${EXTENSIONS.map.build}`;
+    const sourceChecksum = new Source(input, fileName).checksum();
     const cachePath = joinPath(
       Config.instance().cachePath,
       ...[...(info.scope !== undefined ? [info.scope] : []), info.profileName]
     );
-    const hashedPath = joinPath(cachePath, hashedName);
+    const path = joinPath(
+      cachePath,
+      `${info.providerName}${EXTENSIONS.map.build}`
+    );
 
     // If we have it in memory cache, just return it
-    if (this.mapCache[hashedPath] !== undefined) {
-      return this.mapCache[hashedPath];
+    if (
+      this.mapCache[path] !== undefined &&
+      this.mapCache[path].astMetadata.sourceChecksum === sourceChecksum
+    ) {
+      return this.mapCache[path];
     }
 
-    // If we already have parsed map in cache file, load it
-    {
-      const parsedMap = await Parser.loadCached(
-        hashedPath,
-        assertMapDocumentNode,
-        this.mapCache
-      );
-      if (parsedMap !== undefined) {
-        return parsedMap;
-      }
+    // If we already have valid AST in cache file, load it
+    let parsedMap = await Parser.loadCached(
+      path,
+      isMapDocumentNode,
+      this.mapCache,
+      new Source(input, fileName).checksum()
+    );
+    if (parsedMap !== undefined) {
+      return parsedMap;
     }
 
     // If not, delete old parsed maps
-    await Parser.clearFileCache(
-      `${info.providerName}-[0-9a-f]+${EXTENSIONS.map.build}`,
-      cachePath
-    );
+    await Parser.clearFileCache(path);
 
     // And write parsed file to cache
-    const parsedMap = parseMap(new Source(input, fileName));
-    await Parser.writeFileCache(
-      parsedMap,
-      this.mapCache,
-      cachePath,
-      hashedPath
-    );
+    parsedMap = parseMap(new Source(input, fileName));
+    if (!isMapDocumentNode(parsedMap)) {
+      //TODO: more helpful error - can this be product of not matching AST package versions?
+      throw new UnexpectedError('This should not happened');
+    }
+    await Parser.writeFileCache(parsedMap, this.mapCache, cachePath, path);
 
     return parsedMap;
   }
@@ -76,44 +76,50 @@ export class Parser {
       scope?: string;
     }
   ): Promise<ProfileDocumentNode> {
-    const hash = Parser.hash(input);
-    const hashedName = `${info.profileName}-${hash}${EXTENSIONS.profile.build}`;
+    const sourceChecksum = new Source(input, fileName).checksum();
     const cachePath = joinPath(
       Config.instance().cachePath,
       ...[...(info.scope !== undefined ? [info.scope] : [])]
     );
-    const hashedPath = joinPath(cachePath, hashedName);
+    const path = joinPath(
+      cachePath,
+      `${info.profileName}${EXTENSIONS.profile.build}`
+    );
 
     // If we have it in memory cache, just return it
-    if (this.profileCache[hashedPath] !== undefined) {
-      return this.profileCache[hashedPath];
+    if (
+      this.profileCache[path] !== undefined &&
+      this.profileCache[path].astMetadata.sourceChecksum === sourceChecksum
+    ) {
+      return this.profileCache[path];
     }
 
-    // If we already have parsed map in cache file, load it
-    {
-      const parsedProfile = await Parser.loadCached(
-        hashedPath,
-        assertProfileDocumentNode,
-        this.profileCache
-      );
-      if (parsedProfile !== undefined) {
-        return parsedProfile;
-      }
+    // If we already have valid AST in cache file, load it
+    let parsedProfile = await Parser.loadCached(
+      path,
+      isProfileDocumentNode,
+      this.profileCache,
+      sourceChecksum
+    );
+    // If we have cached AST, we can use it.
+    if (parsedProfile !== undefined) {
+      return parsedProfile;
     }
 
     // If not, delete old parsed profiles
-    await Parser.clearFileCache(
-      `${info.profileName}-[0-9a-f]+${EXTENSIONS.profile.build}`,
-      cachePath
-    );
+    await Parser.clearFileCache(path);
 
     // And write parsed file to cache
-    const parsedProfile = parseProfile(new Source(input, fileName));
+    parsedProfile = parseProfile(new Source(input, fileName));
+    if (!isProfileDocumentNode(parsedProfile)) {
+      //TODO: more helpful error - can this be product of not matching AST package versions?
+      throw new UnexpectedError('This should not happened');
+    }
     await this.writeFileCache(
       parsedProfile,
       this.profileCache,
       cachePath,
-      hashedPath
+      path
     );
 
     return parsedProfile;
@@ -123,8 +129,9 @@ export class Parser {
     T extends MapDocumentNode | ProfileDocumentNode
   >(
     path: string,
-    assertion: (node: unknown) => T,
-    cache: Record<string, T>
+    guard: (node: unknown) => node is T,
+    cache: Record<string, T>,
+    sourceHash: string
   ): Promise<T | undefined> {
     let fileExists = false;
     try {
@@ -132,27 +139,30 @@ export class Parser {
     } catch (e) {
       void e;
     }
-    if (fileExists) {
-      const parsed = assertion(
-        JSON.parse(await fsp.readFile(path, { encoding: 'utf8' }))
-      );
-      cache[path] = parsed;
-
-      return parsed;
+    if (!fileExists) {
+      return undefined;
     }
+    const loaded = JSON.parse(
+      await fsp.readFile(path, { encoding: 'utf8' })
+    ) as unknown;
+    //Check if valid type
+    if (!guard(loaded)) {
+      return undefined;
+    }
+    //Check if checksum match
+    if (loaded.astMetadata.sourceChecksum !== sourceHash) {
+      console.log('CHECK');
 
-    return undefined;
+      return undefined;
+    }
+    cache[path] = loaded;
+
+    return loaded;
   }
 
-  private static async clearFileCache(
-    nameFormat: string,
-    path: string
-  ): Promise<void> {
-    const cachedFileRegex = new RegExp(nameFormat);
+  private static async clearFileCache(path: string): Promise<void> {
     try {
-      for (const file of (await fsp.readdir(path)).filter(cachedFile =>
-        cachedFileRegex.test(cachedFile)
-      )) {
+      for (const file of await fsp.readdir(path)) {
         await fsp.unlink(joinPath(path, file));
       }
     } catch (e) {
@@ -176,11 +186,5 @@ export class Parser {
       // Fail silently as the cache is strictly speaking unnecessary
       void e;
     }
-  }
-
-  private static hash(input: string): string {
-    return createHash('shake256', { outputLength: 10 })
-      .update(input, 'utf8')
-      .digest('hex');
   }
 }