+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Removed unnecessary result propagation in interpreter - using stackTop.result instead
 - Terminal outcome statements should now correctly exit from iteration
 - MappedHTTPErrors and MappedErrors are handled a little better
+- Parser uses AST metadata to check if source was changed
 
 ### Fixed
 - operation call no longer overwrites already mapped result